requirements.txt
functions-framework==3.*
markdown2==2.*
pyyaml==6.*
feedparser==6.*
google-genai==1.*
flask